+++ scripts/nodeModuleLoader.js
@@ -3,3 +3,4 @@ const fs = require('fs');
 const sqlite3 = require('sqlite3').verbose();
 const url = require('url');
 const electron = require('electron');
+const semver = require('semver');