+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # [unreleased version]
+### New feature
+- A little memo for every character
+- Auto upgrade database to meet latest features
+
+### Added
+- A version information is now right below the page
+
 ### Improved
 - Displays TA cooldown time at main table.
 - Use ● as a indicator of TA cooldown status instrad of changing background.